Part 2 of immutable Extension refactor.
I made Extension a refcounted object, and privitized the existing
con/destructor and InitFromValue. The only way to get an Extension is to call
a factory method.

In the next CL, I plan to make the factory method return a const Extension,
to guarantee that no one can modify the Extension object after creation.

Note: There was a tricky part of this CL because of the difference in
semantics between scoped_ptr and scoped_refptr. I had to be careful not to use
ptr.release(), since that would result in leaks (an un-Released AddRef).

BUG=56558
TEST=no functional change

Review URL: http://codereview.chromium.org/3982001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@63919 4ff67af0-8c30-449e-8e8b-ad334ec8d88c