Re-enable ExtensionApiTest.WindowOpen.
There was a bug in the original implementation of this test
that caused it to close early. Depending on the exact
timing, this sometimes failed, which caused it to be marked
flaky. Later, a bug was introduced that broke the underlying
functionality, which is also fixed here.

BUG=60156
TEST=Covered by browser_tests.

Review URL: http://codereview.chromium.org/4026005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@63949 4ff67af0-8c30-449e-8e8b-ad334ec8d88c