Attempt 2 at landing top sites refactor: Refactors TopSites so that it's hopefully easier to maintain and doesn't suffer the plethora of threading issues that exist with the current version. Here's the breakdown of what was refactored:
. TopSitesCache: Contains the most visited urls and thumbnails.
. TopSitesBackend: All mutations to topsites data end up calling into
the backend on the UI thread. TopSitesBackend processes the method on
the DB thread calling through to the TopSitesDatabase.
. TopSites: uses two TopSitesCache. One that contains the raw history
data, the other contains the processed data (pinned/blacklisted). The
processed cache can be accessed on any thread.

TopSites waits until history loads to know if it should migrate or use
it's own db. I could probably make these execute in parallel, but for
now this is how it works.

This patch also makes it so the dom ui accesses the thumbnails on the
IO thread.

BUG=56382
TEST=Make sure all your thumbnails are correctly updated and you don't
see problems.

Review URL: http://codereview.chromium.org/4051004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@64072 4ff67af0-8c30-449e-8e8b-ad334ec8d88c