Restrict total parallel DNS resolutions
A/B tests suggest that large numbers of parallel
resolutions may cause the DNS failure rate
(re: ratio of "host not found" to "found") to
rise.  To stay safely away from that threshold,
this change lowers the maximum parallel resolutions
to 8, and restricts the speculative resolutions
to 3.

We are also running A/B tests which will look at the
impact of modulating either of these values.

BUG=3041
r=eroman
Review URL: http://codereview.chromium.org/4111004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@64369 4ff67af0-8c30-449e-8e8b-ad334ec8d88c