Cleanup histogram syncronizer a bit
I adjusted the code so tha there was no chance
for the pending response count to get to zero
until all the increments had been applied, and
all the rendereres had been notified.  Without
this change, there was a chance that the tally
would hit zero prematurely, triggering a
a premature completion.

I went through the code clarifying where
a lock was needed, and where it wasn't.
In all cases, we now either take a lock, or
assert we're consistently on a single
thread (meaning we don't need a lock).

r=davemoore
Review URL: http://codereview.chromium.org/4193008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@64781 4ff67af0-8c30-449e-8e8b-ad334ec8d88c