[tabbed options] use CSS to order confirm/cancel dialog buttons
This adds an action-area class which handles positioning and padding (anchors the action strip to the bottom right), which is basically what the old button-strip did). Now, button-strip just handles ordering of the buttons, and on Windows it is reversed. One usage is to create a div with both button-strip and action-area classes on it, which is convenient if the only thing in the action area is buttons. The other usage is to create an action-area div, and put children in it, one of which is a button-strip.

For now, views is [OK] [Cancel] and all others are [Cancel] [Ok]. I'll ping zelidrag and alcor about what we want chromeos to be.

BUG=none
TEST=manual

Review URL: http://codereview.chromium.org/4515001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@65149 4ff67af0-8c30-449e-8e8b-ad334ec8d88c