Fix dbus- and geolocation- related crash on Linux.
The NetworkManagerWlanApi class in wifi_data_provider_linux.cc was
using a shared dbus connection which was referenced as a GSource in
the glib message loop in the main thread, and also from the
WifiDataProviderCommon thread. When the dbus proxy object and connection
was unreferenced in the WifiDataCommon thread, the glib main loop in the
main thread could concurrently reference it during object destruction,
causing a crash.

We fix this by using a private dbus connection instead, referencing its
own glib main loop separate to that of the main thread.

BUG=59913
TEST=GeolocationWifiDataProviderPlatformTest.CreateAndDestroy

Review URL: http://codereview.chromium.org/4307001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@65185 4ff67af0-8c30-449e-8e8b-ad334ec8d88c