Changing the installer switches from wchar_t[] to char[].
Because of this I'm also refactoring some code that before
was using wstring to build command lines by hand instead of
using the CommandLine class.  Now we use CommandLine.

To get this to work correctly, I also needed to fix CommandLine::AppendArguments so I added a little test for it.

TEST=There should be no changes in functionality.  Run all installer tests.
BUG=61609

Review URL: http://codereview.chromium.org/4928002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@66088 4ff67af0-8c30-449e-8e8b-ad334ec8d88c