Collect DirectX diagnostic information asynchronously.
Collecting this can take a couple of seconds. I put the code onto a worker thread. The about:gpuinfo handler polls for it until it is available, initially displaying only the subset of informtation that can be retreived quickly.

This makes the startup time for accelerated compositing, WebGL, etc significantly lower on Windows. None of the other platforms have this issue.

TEST=go to about:gpuinfo, try
BUG=59711

Review URL: http://codereview.chromium.org/4860001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@66184 4ff67af0-8c30-449e-8e8b-ad334ec8d88c