Implement a dispatcher for prerender requests
This CL provides a dispatcher that can feed the prerender system with
HTML type requests that were prefetched.

One thing is missing though: this sends all requests in, and doesn't tag the
origin request.  Unfortunately referer information is often tossed out in
resource_dispatcher_host...

Is there a private place it's sent along to for this type of use?  I left that
undone right now but I'm open to suggestions about a good way to do it.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/4655004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@66790 4ff67af0-8c30-449e-8e8b-ad334ec8d88c