[Mac] Fix a memory corruption when closing a JS modal dialog.
JSModalDialogCocoa::CloseAppModalDialog was calling [NSApp endSheet],
which in turn was deleting the JSModalDialogCocoa (via a scoped_ptr
in alertDidEnd).

After that, the CloseAppModalDialog was writing NULL to the alert_
member variable, but at that point the object was deleted,
so it was corrupting memory.

GuardMalloc detected this corruption very quickly.

It was also causing BrowserTest.JavascriptAlertActivatesTab to crash
flakily on Mac. This change undisables that test.

The bug contains more details, context, and stack traces.

BUG=54896
TEST=browser_tests

Review URL: http://codereview.chromium.org/5318007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@67493 4ff67af0-8c30-449e-8e8b-ad334ec8d88c