This CL add a GetInstance() method to singleton classes instead of relying on the callers to use Singleton<T>.
In some cases I have used the LazyInstance<T> pattern as that was simpler.

This is a small step towards making all singleton classes use the Singleton<T> pattern within their code and not expect the callers to know about it.

I have selected all files under src/app and src/base which use Singleton<T> in this CL. Once this CL goes in I'll work on the rest of the files.

BUG=65298
TEST=all existing tests should continue to pass.

Review URL: http://codereview.chromium.org/5527004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@68300 4ff67af0-8c30-449e-8e8b-ad334ec8d88c