Fix handling of 'Use system proxy settings' policy
The problem is that using system proxy settings is triggered when the
internal proxy prefs of Chrome are set to their default values.
Unfortunately it is possible that these prefs are set to their defaults but
they are not controlled by the default pref store. This CL fixes this by
checking for the default values and not only for the origin of prefs. A longer term solution is described in
http://crbug.com/65732

BUG=65736
TEST=manual

Review URL: http://codereview.chromium.org/5664001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@68591 4ff67af0-8c30-449e-8e8b-ad334ec8d88c