Create a new container to own downloads in DownloadManager.
Shifts destructor to use new container, and removes no longer needed containers.  Part of the downloads refactor effort.

dangerous_finished_ was only used for finding downloads that needed to be removed on disk, which can be determined more cleanly from the DownloadItem state.  This does mean two passes through downloads_ (one to find the downloads that require special handling, and one to delete everything) but a) it's worth it for code clarity, and b) when we cleanup the semantics of DownloadItem::Remove, we can drop it to a single pass.

BUG=63493
TEST=All known not-always failing download tests pass.

Review URL: http://codereview.chromium.org/5636003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@68877 4ff67af0-8c30-449e-8e8b-ad334ec8d88c