Revert "Add named testing interface."
http://src.chromium.org/viewvc/chrome?view=rev&revision=67300

The change has been approved conditionally,
i.e. if the outstanding review comments are handled
in a follow-up as the top priority.

Two weeks have passed, and I have not seen the follow-up.
It is risky to keep this change in the tree, as it is already
non-trivial to revert.

Conflicts:

	chrome/browser/automation/automation_provider.cc
	chrome/test/ui/named_interface_uitest.cc
	ipc/ipc_channel_posix.cc

TBR=nirnimesh

BUG=chromium-os:8512
TEST=none

Review URL: http://codereview.chromium.org/5707006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@68944 4ff67af0-8c30-449e-8e8b-ad334ec8d88c