Clean up PluginDataRemover.
This CL fixes some threading issues in PluginDataRemover. It also adds a static method to find out if removing plugin data is supported by an installed plugin, and a method to set the completion task while it is in the process of removing data.

BUG=58235
TEST=none

Review URL: http://codereview.chromium.org/5715001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@69221 4ff67af0-8c30-449e-8e8b-ad334ec8d88c