Prevent App extensions to be returned by AutomationProvider::GetEnabledExtensions().
These are not useful for testing and are causing CEEE problems, since it expects a single extension being installed.

A better fix might be to not have any default extension installed with Chromium and only have them with a branded Chrome so that they don't show up in Chrome Frame or CEEE installs... Maybe... Started a conversation with the Chrome Apps and Extensions team to discuss that.

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/6008004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@69867 4ff67af0-8c30-449e-8e8b-ad334ec8d88c