Introduce a separate preference for 'proxy server mode'
The new preference is kProxyServerMode, which supersedes kProxyAutoDetect and kNoProxyServer. The point of this change is to represent 'use system proxy settings' in a more robust way. The proxy extension API is also adjusted to the preference system.

This is a continuation of gfeher's patch from issue 5701003.

BUG=65732, 66023
TEST=ProxyPrefsTest.*, and also covered by ExtensionApiTest.Porxy*, PrefProxyConfigServiceTest.*

Review URL: http://codereview.chromium.org/6004003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@70042 4ff67af0-8c30-449e-8e8b-ad334ec8d88c