Mac: Don't skip words when user clicks ignore in the spelling panel
Clicking Ignore in the spelling panel would skip misspelled words.

The problem was that the Chromium code and the spell checker were both advancing to the next word. This caused us to advance twice instead of once.

Fix was to simply remove the Chromium code that was advancing to the next word.

BUG=68557
TEST=Downloaded version 9.0.597.19 of Google Chrome and verified that the bug existed on both 10.5 and 10.6. Applied my patch and verified that pressing ignore worked correctly on both 10.5 and 10.6.

Review URL: http://codereview.chromium.org/6103001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@70532 4ff67af0-8c30-449e-8e8b-ad334ec8d88c