Refactored app cache clear on exit code to happen in the object owning the files.
In an effort to remove the static functions that used to be called from the
BrowserProcessImpl very late in the shutdown process and move the code to
immediately after the files get freed from their respective users. Which
will help to parallelize the shutdown sequence better and ensure files are
deleted when they are not used anymore.

BUG=65076
TEST=TBA

Review URL: http://codereview.chromium.org/6077005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@70893 4ff67af0-8c30-449e-8e8b-ad334ec8d88c