Fix a possible crash in the ClientSideDetectionService.
This fixes a problem where if the model fetch is still running when the
ClientSideDetectionService is destroyed, the URLFetcher is not deleted.
This could potentially cause the OnURLFetchComplete callback to be called
on a deleted ClientSideDetectionService object.

Now the model fetcher will always be deleted, and we will also make sure to
destroy the ClientSideDetectionService before tearing down the IO thread
(since the URLFetcher dtor can post messages there).

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/6298004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@71901 4ff67af0-8c30-449e-8e8b-ad334ec8d88c