Force patches that touch sync code to use the sync trybots
As of today, we have dedicated chromium trybots that can run the sync
integration tests, but it requires developers to explicitly send try
jobs to these bots via the --bot switch.

Now that the sync trybots are stable, it is time to make this process
automatic by adding a PRESUBMIT.py file to chrome/browser/sync, that
adds the sync trybots in addition to win, mac and linux to the list of
preferred trybots.

This patch adds such a PRESUBMIT.py file, that includes
win_sync, mac_sync and linux_sync to the list of preferred trybots.

BUG=70311
TEST=modify something under chrome/browser/sync and do a "git try" or "gcl try"

Review URL: http://codereview.chromium.org/6275010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@72067 4ff67af0-8c30-449e-8e8b-ad334ec8d88c