chromeos: Refresh static about:os-credits page.
This is the result of running my scripts against a new build
of x86-mario's packages in a clean sysroot.  Not much
changed.

We got updated (possibly unchanged) licenses due to
upgrading to newer versions of the following packages:

  app-admin/sudo
  app-i18n/ibus
  app-i18n/ibus-m17n
  app-i18n/ibus-pinyin
  dev-cpp/glog
  dev-libs/libgcrypt
  dev-libs/libpcre
  sci-geosciences/gpsd
  sys-apps/mosys
  sys-power/powertop

We dropped the following packages:

  app-text/iso-codes
  dev-libs/libcroco
  gnome-base/librsvg
  gnome-extra/libgsf
  x11-libs/libxklavier

We added the following packages:

  sys-libs/talloc
  x11-apps/xtrace
  x11-terms/rxvt-unicode

We picked up the tarball's license from:

  dev-libs/glib

I do not understand the appearance of xtrace; I only added
it to chromeos-dev and I did a non-dev build here.
Whatever.

BUG=none
TEST=looked at it
TBR=davemoore

Review URL: http://codereview.chromium.org/6357013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@72347 4ff67af0-8c30-449e-8e8b-ad334ec8d88c