Fix uninitialized variables introduced in CL 72374.
time_slice_hrs is being used uninitialized at line 459, and this was causing a Run-Time Check failure and crash in Debug versions on Windows.

I don't know whether 0 is the the correct value in this case, but this fixes the crash. mirandac@ will need to review the behavior.

BUG=None
TEST=Run debug version of Chrome on Windows and wait a few seconds.
Review URL: http://codereview.chromium.org/6385004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@72444 4ff67af0-8c30-449e-8e8b-ad334ec8d88c