Fix a bug where we were showing a wrong language name on the OOBE screen.
The OOBE screen is by defaut shown in English, but is also shown in a
different language if "initial_locale" value is set in
/mnt/partner_partition/etc/startup_manifest.json

In this case, the default language name of the language selection menu
should be the one specified by "initial_locale". However, we were always
showing English before this fix.

BUG=chromium-os:11345
TEST=See the instruction at crosbug.com/11345

Review URL: http://codereview.chromium.org/6307015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@72762 4ff67af0-8c30-449e-8e8b-ad334ec8d88c