Simplified PrefValueStore::NotifyPrefChanged.
Reason for the refactoring is the need to properly handle disappearing policies.
This situation arises when a managed preference is removed that only has
been registered in one pref store (for example local_store) the notification
is sent to all pref stores though but the profile pref store doesn't have
a default value for that pref and therefore ControllingPrefStoreForPref returns
INVALID_STORE which shouldn't be treated as an exceptional situation (DCHECKed).

BUG=71028
TEST=unit_tests --gtest_filter=*Pref* should give pretty good coverage.

Review URL: http://codereview.chromium.org/6357019

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@72787 4ff67af0-8c30-449e-8e8b-ad334ec8d88c