Refactoring to use a new IPC message filter.
There were several ViewHostMsgs being sent to the GpuProcessHost via a RenderMessageFilter. This patch creates a GpuMessageFilter and renames the ViewHostMsgs to GpuHostMsgs. This will clean up some of the clutter in render_messages_internal and render_message_filter. It will also allow us to easily add new messages to the GpuProcessHost in the future.

BUG=none
TEST=by hand and trybots

Review URL: http://codereview.chromium.org/6401003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@73152 4ff67af0-8c30-449e-8e8b-ad334ec8d88c