Update toolbar for killed tabs
Switching to a killed tab wouldn't update the toolbar.

This was a regression introduced by r71360. We were checking to see switches::kReloadKilledTabs was set and if so we were reloading the tab and skipping the update.

Unfortunately we were also skipping the update if switches::kReloadKilledTabs was not set. Fix was to only skip the update if switches::kReloadKilledTabs was set.

BUG=70306
TEST=Reproduced the bug. Applied my fix. Verified that the toolbar correctly updates for killed tabs.
Review URL: http://codereview.chromium.org/6312102

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@73534 4ff67af0-8c30-449e-8e8b-ad334ec8d88c