Refactoring GpuProcessHost for use by the browser.
This is a step towards allowing the browser to make requests of the GpuProcessHost (it's traditionally done by renderers). This is done with a level of indirection. It replaces the sending of IPCs to the renderer with calling a callback (which now sends the IPCs). The IPCs sent from callbacks originating in GpuMessageFilter use WeakPtrs to guard against the filter disappearing before the callback is called.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/6287026

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@73624 4ff67af0-8c30-449e-8e8b-ad334ec8d88c