Create a ProfileIOData object to own ChromeURLRequestContext members.
This is the beginning of the ProfileIOData object. I hope to move more and more ChromeURLRequestContext members into it, so we can stop refcounting the members. Originally, I wanted ProfileIOData to own the ChromeURLRequestContext objects, but it would have caused too many difficulties with lifetime management for now. I'll save that for the future. This is an improvement since now URLRequestContext's members don't always have to be refcounted.

TODO: Create an analog for OffTheRecordProfile objects.

BUG=67237
TEST=existing

Review URL: http://codereview.chromium.org/6315013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@73651 4ff67af0-8c30-449e-8e8b-ad334ec8d88c