Only call NotifyAndTerminate when we've explicitly signed out.
NotifyAndTerminate shuts down the login and session manager, which we do not want to or need 
to do in other cases. This specifically fixes the case where browser shuts down when all the 
windows are gone (as in pyauto) or a kill event is received (other python tests).
BUG=chromium-os:11474
TEST=This should fix python tests on the device.

Review URL: http://codereview.chromium.org/6410066

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@73824 4ff67af0-8c30-449e-8e8b-ad334ec8d88c