FilePath: Remove much of ToWStringHack, adding a LossyDisplayName()
The reason we don't want a free conversion between FilePaths and Unicode
is that it can be lossy.  But when displaying a string to the user,
we're ok if it's lossy when we have no other option.

This change introduces a LossyDisplayName() method that returns a
string16, and converts many of the users of ToWStringHack to use it.

BUG=69467

Review URL: http://codereview.chromium.org/6246036

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@73840 4ff67af0-8c30-449e-8e8b-ad334ec8d88c