Add debug logging for mixed scripting.
Our mixed scripting can be very difficult to debug when the offending
request comes from nested iframes and the like because the issue won't
show up in the Javascript console for the page. This also occurs if a
redirect page loads mixed scripting.

This change cleans up a previous change (r71881) now that the WebKit
side of this has landed, and logs mixed scripting issues to the debug
log (when logging is enabled).

BUG=none
TEST=Run with --enable-logging and go to a site with mixed scripting. Confirm that a debug entry appears in the chrome_debug.log.

Review URL: http://codereview.chromium.org/6246149

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@74119 4ff67af0-8c30-449e-8e8b-ad334ec8d88c