Remove SQLite safe-browsing backing store.
In October 2010, 7.0 rolled to stable with the safe-browsing database
configured to read-from-sqlite and write-to-new-file-format.  There is
still a long tail of SQLite-format databases in the wild, but very
small (and possibly due to a systemic issue with updating, so it may
not get smaller).

This change flips the switch to no longer support reading the SQLite
files.  Those users will do a full re-sync with the safe-browsing
servers.

BUG=58552
TEST=none

Review URL: http://codereview.chromium.org/6413022

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@74253 4ff67af0-8c30-449e-8e8b-ad334ec8d88c