Reworks autocomplete result set processing by the edit/model necessitated by my last change (only one result set vs 2). There is now one notification when the results change. The edit is the sole listener and updates the popup after it does some processing.
My last change broke keyword state when switching tabs. This is because of the change I did to AutocompleteEditModel::SetUserText, specifically:

  keyword_.clear();
  is_keyword_hint_ = false;

When you restore a tab we end up in

void AutocompleteEditModel::RestoreState(const State& state) {
  ...
    keyword_ = state.keyword;
    is_keyword_hint_ = state.is_keyword_hint;
    view_->SetUserText(state.user_text,
        DisplayTextFromUserText(state.user_text), false);

Which ends up back in SetUserText and we because of my change the
keyword is effectively chucked.

I added the code to clear out the keyword because of the change of
notifications. Specifically because AutocompletePopupModel::Observe is
now notified before AutocompleteEditModel::Observe yet it expects the
AutocompleteEditModel to have been updated.

Additionally I need to update the keyword_ and keyword_hint_ immediately in the edit if the results go empty. This mirrors what the old code was doing.

BUG=72022
TEST=covered by tests

Review URL: http://codereview.chromium.org/6484015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@74823 4ff67af0-8c30-449e-8e8b-ad334ec8d88c