Use views::View::GetPreferredSize instead of BaseTabStrip::GetPreferredSize
We call GetPreferredSize on the tabstrip object, but we were doing so using the BaseTabStrip interface rather than the more general views::View interface.  This CL changes it so that we refer to the tabstrip object by its views::View interface.

By asking the container to plan its layout given an arbitrary view rather than a specific BaseTabStrip, the code is more general and less dependent on changes in the BaseTabStrip class.  I just think it's more tidy to have containers know as little as possible about of the intimate details about their contained items.  It's cleaner and simpler.  We don't need to know about the BaseTabStrip class in the FrameView class.  We don't need to include the extra header.

Follows 6312196.

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/6465025
Patch from Chad Faragher <wyck@chromium.org>.

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@74832 4ff67af0-8c30-449e-8e8b-ad334ec8d88c