Make the use of WeakPtr threadsafe.
This checks that the WeakPtrs to GpuMessageFilter are always accessed from the same thread. It also ensure that the GpuMessageFilter is destroyed from that thread. This fixes a potential race where the GpuMessageFilter was being destroyed (on the IO thread) but a callback was trying to access it (on the UI thread).

BUG=none
TEST=stepped through ~GpuMessageFilter to ensure that switching it to the UI thread was safe

Review URL: http://codereview.chromium.org/6530004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@75160 4ff67af0-8c30-449e-8e8b-ad334ec8d88c