Fixes the context menu "save as" when policy is effective.
Added a solution for the save as link still trying to open "Save As"
dialog instead of using the policy location when set. We still have to
keep this option in the menu enabled so that people can work around automatic
open which is not dependent on this policy.

BUG=59768
TEST=Set the policy. Right click a download link and "save as" should save without prompting.

Review URL: http://codereview.chromium.org/6546047

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@75587 4ff67af0-8c30-449e-8e8b-ad334ec8d88c