[Reason for revert: might be a factor in XP-only failures UITestCanLaunchWithOSMesa.  Will unrevert if proven innocent.]
Revert 76075 - With this CL, GPU blacklist auto update from the web (before it ships together with chrome) is implemented.  This allows us to blacklist bad GPU/drivers as soon as we discover them.

Note that this patch does not turn the auto update on.  We will turn it on in a separate CL.

BUG=68802
TEST=build bots all green

Review URL: http://codereview.chromium.org/6469094

TBR=zmo@google.com
Review URL: http://codereview.chromium.org/6599020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@76122 4ff67af0-8c30-449e-8e8b-ad334ec8d88c