Initial landing of file_manager component extension (not yet enabled)
This is the initial landing of a component extension to handle file open, save-as, and file management user interfaces.  This is just the basic plumbing, minus any useful functionality.

I've tested this by building with...
$ export GYP_DEFINES='file_manager_extension=1'
$ make chrome -j16

Then launch the result and visit chrome-extension://hhaomjibdihmijegdhdafkllkbggdgoj/main.html, and verify that the <h1> tab is blue, and the document contains "Hello World".

BUG=chromium-os:11988
TEST=

Review URL: http://codereview.chromium.org/6588053

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@76408 4ff67af0-8c30-449e-8e8b-ad334ec8d88c