Don't run the phishing classifier if the page was loaded from a private IP address.
Since we can't discern the remote IP address if the page was loaded via a proxy, also disable phishing detection in that case.

The pre-classification checks are moved to ClientSideDetectionHost, since it is already set up as a TabContentsObserver and receives the ViewHostMsg_FrameNavigate_Params for each navigation.

BUG=none
TEST=ClientSideDetectionHostTest.ShouldClassifyUrl

Review URL: http://codereview.chromium.org/6594006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@76660 4ff67af0-8c30-449e-8e8b-ad334ec8d88c