This is the first CL for a sequence of CLs that will add in the display infrastructure that I have built for the DOM Login and touch keyboard. If have created a reference CL for what the end product of this sequence is intended to be, so that I can give context for the CL under review.This can be found at: http://codereview.chromium.org/6577003/
This CL adds in chrome://login-container to touchui==1 and chromeos==1 builds. This page is hooked into the WebUI to enable testing and development of a Browser based container for the DOM Login screens. In this CL all it does it
create a Browser with chrome://login in it. This functionality will be extended in following CLs.

Patch from Ryan Harrison <rharrison@chromium.org>

BUG=none
TEST=manual

Review URL: http://codereview.chromium.org/6579003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@76894 4ff67af0-8c30-449e-8e8b-ad334ec8d88c