Refactor GetCommandLineArgumentsCommon to not use strings.
Passing around command lines as strings is dangerous, because a given
command line is interpretable only with respect to a given parser.
This changes a function to instead pass a CommandLine object for
carrying the arguments around, and relies on the callers to stringify
that as necessary.

It turns out that the Linux desktop file format has a well-defined
syntax that is different than the syntax we were using.

BUG=69467

Review URL: http://codereview.chromium.org/6624072

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@77338 4ff67af0-8c30-449e-8e8b-ad334ec8d88c