Initial integration of DOM login screen with existing infrastructure.
This CL integrates the work that has been done on the DOM based login screen with the work on refactoring the login infrastructure. This is a very preliminary CL and is intended to review the logistics of integration, so there are many things with the DOM screen itself are broken. The intent is to address any issues with the seam between the DOM login code and existing infrastructure. Issues related to functionality of the DOM login screen or integration with chromiumos will be addressed in following CLs. This CL is only included in touchui==1 and chromeos==1 builds and NOT the default login screen.

This CL has been tested on the desktop by adding --login-manger --login_screen=login --dom-login --stub-cros to cause chromium to come up with the login screen.You may need to use --login-profile also. It is known to be broken in chromiumos, due to integration issues with the window manager. This is currently being investigated.

I have added in a command line flag "--dom-login" that allows for using the DOM based login screen instead of the views based one. This screen is not very functional at the moment. The WebUI handling layer needed to be rewritten, so we are no longer using AuthenticatorFacade. There is an issue with commands not making it down to the login infrastucture layer at the moment, so shutting down and logging in off the record hang.

The support for chrome://login-container has been removed, because the login page's DOM no longer workers as a standalone window since it needs the login infrastructure. In addition you can access the login screen for development on the desktop as mentioned before, so it isn't needed.

Patch from Ryan Harrison <rharrison@chromium.org>

BUG=none
TEST=manual


Review URL: http://codereview.chromium.org/6729018

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@79993 4ff67af0-8c30-449e-8e8b-ad334ec8d88c