Fix leak in ProfileMenuButtonTest.MenuTest
Valgrind was reporting a leak in -[ProfileMenuDelegate menuWillOpen:].

I wasn't able to reproduce this on my machine with a Debug or Release build.

As a workaround I'm simply removing that part of the test. The test now just checks that popUpContextMenu:withEvent:forView: is called. This is still pretty good since it tests the mouse down code and the pressed state drawing code.

BUG=77910
TEST=Wasn't able to reproduce the same leak but I did get some other leaks in the menu code. Applied my patch and verified that valgrind ran with no errors.

Review URL: http://codereview.chromium.org/6677169

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@80687 4ff67af0-8c30-449e-8e8b-ad334ec8d88c