Make ExtensionSpecialStoragePolicy hold references to extentions added.
The 'policy' object can outlive the ExtensionService class and upon destruction the ExtensionService class does not clear out the 'policy' object. Prior to the change, the 'policy' class was keeping raw ptrs to 'extension' objects, but they will have been deleted if the ExtensionService has been deleted... and CRASH!

With the change, the 'policy' object addrefs each extension added to it... so no crash.

BUG=78688
TEST=existing tests apply
Review URL: http://codereview.chromium.org/6801062

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@80971 4ff67af0-8c30-449e-8e8b-ad334ec8d88c