File Manager: synthesize root directory contents
The local filesystem does not support enumeration of the root directory.  This patch works around the issue by telling the file manager the contents of the root directory ahead of time.  A side-effect "bonus" is that we can localize the UI for these directory names without actually changing them on disk.

When the dialog is loaded in the test harness, the list of root entries is taken directly from the filesystem to make things easier on the developer (otherwise they'd have to prime the test filesystem in a particular way.)

BUG=chromium-os:13845
TEST=manual testing in harness and in chromeos-chrome

Review URL: http://codereview.chromium.org/6840006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@81439 4ff67af0-8c30-449e-8e8b-ad334ec8d88c