Refresh cache in UserCrosSettingsProvider after storing policy
This is necessary because UserCrosSettingsProvider has no way of
noticing that device settings have changed and thus cannot update its
local cache. In order to work around this, we trigger a explicit reload
so the caches get updated. This is only temporary and should go away
when we do the big SignedSettings interface refactoring on the Chrome
side.

BUG=chromium-os:14188
TEST=manual

Review URL: http://codereview.chromium.org/6864010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@81777 4ff67af0-8c30-449e-8e8b-ad334ec8d88c