Add histograms that measure unpack time.
These histograms should allow us to decide if a UI for unpack progress will be necessary.  We will have to extrapolate for small extensions to larger ones.

For example, if extensions in the 1-2 MB range take a second to unpack, and 2-3 MB takes two seconds, then we can guess that 100MB extensions will take over a minute, so a UI is clearly needed.

BUG=None
TEST=Manual: Install a few extensions, look at about:histogram

Review URL: http://codereview.chromium.org/6875008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@82289 4ff67af0-8c30-449e-8e8b-ad334ec8d88c