[Mac] Show poof animation for current window only when deleting BMB bookmarks
Checks |isKeyWindow| before performing the poof animation. Re-ordered the
call to [oldButton removeFromSuperview] so that the view will actually have
a window at the time of the check. Also, moved initialization of |poofPoint|
inside the block to avoid doing the work when it's not needed.

BUG=80046
TEST=See bug.

Review URL: http://codereview.chromium.org/6883121

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@82522 4ff67af0-8c30-449e-8e8b-ad334ec8d88c