Fix CaptureVisibleTabRace to complete deterministically.
Since the test appears to time out on Leopard, leave in a FLAKY_ prefix for
the Mac, and add extra logging to see how far it gets on the bots.

Also fix expected/actual argument order in chrome.test.assertEq call and
the way we report kInternalVisibleTabCaptureError (we assign it to error_
directly in all places but one).

BUG=80212
TEST=ExtensionApiTest.CaptureVisibleTabRace
Review URL: http://codereview.chromium.org/6904054

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@83215 4ff67af0-8c30-449e-8e8b-ad334ec8d88c