Move UserMetrics to content.
Just the class that provides the action logging API which is used all over the place
is being moved. The UserMetrics class uses the notification system to inform the core
of the  user metrics system of each event, that core part is not being moved.

Also take care of a TODO to remove the class methods that take a Profile*
as a paramter (since Profile is verbotten in /content).

BUG=78499
Review URL: http://codereview.chromium.org/6883021

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@83251 4ff67af0-8c30-449e-8e8b-ad334ec8d88c