With this implementation, logging on to sync logs you in to the multiprofile system (if --multi-profiles flag is on).
A drop-down menu then appears, from which you can create a new profile; this will open a browser window from which you can sync to a different Google account.

When the browser is shut down, the active profile is saved. This profile will be used when the browser is reopened.

Many things need to be added -- there's no duplicate checking, no way to delete a profile, and no color scheme -- etc.  This is just the most basic multi-profile functionality.

BUG=60105
TEST=existing tests; full suite to come.
Review URL: http://codereview.chromium.org/6881054

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@83261 4ff67af0-8c30-449e-8e8b-ad334ec8d88c