Clean up ExtensionService shutdown and related notifications.
It is not necessary to unload extensions during destruction because that only happens when the profile is going away, so the unloading would be wasted effort.

With that gone, you can also get rid of DestroyingProfile(), which simplifies things.

Also moved all hooking of outside components that happens during extension load and unload into a central location.

Review URL: http://codereview.chromium.org/6902059

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@83408 4ff67af0-8c30-449e-8e8b-ad334ec8d88c