ntp4: fix most visited blacklisting
Added a new function refreshData that replaces the current most visited page list with the new most visited page list, but tries to minimize turbulence, such that, for example, blacklisting a page will just replace that page without reordering everything else.

This is slightly more complicated than the old ntp's implementation, which just overrode the next setMostVisitedPages callback when a page was blacklisted. But that seems prone to race conditions, and it also doesn't work that great for non-blacklist updates.

BUG=none
TEST=included

Review URL: http://codereview.chromium.org/6901035

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@83743 4ff67af0-8c30-449e-8e8b-ad334ec8d88c