Mac: Use bitmaps to draw toolbar background
This is a part of the general change to remove the use of programatically drawn controls and use bitmaps instead.

With this change we now draw the toolbar and download shelf using bitmaps. Note, we didn't have a pre-existing bitmap for inactive toolbars so I create one myself. I'll ping Cole for a better one later.

While I was in this area of code I made some other changes as well:
- added a white highlight at the top of the download shelf
- change the look of the download shelf a bit. Previously the download shelf looked curved (light -> drak -> light). I offset the patter a bit to give a simpler light -> dark look instead. This matches the Windows look as well.
- fill the background using NSRectFillUsingOperation(rect, NSCompositeSourceOver) to properly blend semi-transparent theme images with the background.

Mac Screenshots:
http://www.dropmocks.com/mUnyO
Window screenshot for comparison:
http://www.dropmocks.com/mUni5

BUG=53217
TEST=Ran with and with out themes. Verified that things looked good.

Review URL: http://codereview.chromium.org/6931002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@84304 4ff67af0-8c30-449e-8e8b-ad334ec8d88c