Fix "chrome://bookmarks" URL not displaying in omnibox.
This change modifies the BrowserURLHandler to handle reverse mapping of the chrome-extension://.../ bookmark manager URL back to chrome://bookmarks.

To do this, I've added functionality to the BrowserURLHandler to support reverse mapping even when there is no original URL. The previous system would only reverse remap if it was mapped earlier in the other direction from what the user had typed into the omnibox. This is not the case when the bookmark manager updates its URL by adding refs (#1, #12, etc) at the end of the URL string - which loses the original URL typed by the user.

To support the above scenario, I've added new semantics to the BrowserURLHandler so that when the first handler (forward mapper) is NULL, it always tries the reverse remapper which should do the reverse mapping only if the URL matches one of its expectations. I've added ExtensionWebUI::HandleChromeURLOverrideReverse(), which is such a reverse mapper.

Additionally, I've made BrowserURLHandler a singleton instead of a class with static methods for easier testing and added a couple of tests for it.

Finally, there's a couple of other pieces of plumbing to make this work.

BUG=37356
TEST=Go to the bookmark manager. chrome://bookmarks/ should be shown in the omnibox. Select a folder - this should change the URL to something like chrome://bookmarks/#5. Copy this URL and paste it back into the omnibox. You should be taken to the bookmarks manager with the same selection.
Review URL: http://codereview.chromium.org/6904167

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@84464 4ff67af0-8c30-449e-8e8b-ad334ec8d88c