A hack for whilte flash issue in screen locker.  I tried following but none of them worked. From the log, I do see that wiget gets painted  before FREEZE_UPDATES is sent.
* Paint on map
* Don't send WM_IPC_WINDOW_LOGIN_BACKGROUND with did_paint=1 when background is painted.
* move FREEZE_UPDATE after all child Widgets gets painted
* move FREEZE_UPDATE to next event style using PostTask

 One strange thing is that it is the toplevel window (widget_ in WidgetGtk) that is
flashing, but not window_contents_ that actually paints the contents. (tested by setting window's background)
Sicne widget_ is covered by window_contents_, it should not be visible.
May be right solution is to eliminate a gdkwindow that window_contents_ has, but
it may introduce a lot of regression and may take time to fix them. Since we're removing gtk, I decided to
do this hack instead.

BUG=chromium-os:13154
TEST=lock screen on device. No white flashing anymore.

Review URL: http://codereview.chromium.org/6956004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@84501 4ff67af0-8c30-449e-8e8b-ad334ec8d88c