[keyboard] Do not send Shift modifier for any keys
Previously, we were sending the shift modifier with the key event for
any keys on the shifted layout.  This had the side-effect that the
backend converted "/" to "?" because of the shift modifier.

We don't need to send shift, because we already send capital letters.
Bryan Yeung says that the shift flag was necessary before, but that got
fixed in the backend a while ago.

BUG=
TEST=Have not tested


Review URL: http://codereview.chromium.org/7004003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@84851 4ff67af0-8c30-449e-8e8b-ad334ec8d88c