chromeos: refetch device properties when cellular modem powers on.
When the system starts up, chrome fetches the device properties for
all network devices. However, for cellular devices, many of the
Cellular.XXX properties don't become defined until the modem is
enabled. Since chrome doesn't refetch the properties, nor are any
property change signals sent out from flimflam for these properties
(because generally speaking they are ones that don't change),
chrome's cached copies are empty. The result is that when the Device
tab is brought up in the internet settings page, all the fields are
blank.

Chrome now refetches device properties when it sees the value of
the Powered property change.

Also, I removed the LastUpdate item from the device tab. That property
hasn't existed in flimflam for several months.

BUG=chromium-os:14180
TEST=To demonstrate the problem without this fix, reboot the machine,
go to internet settings, click "Options" for the cellular service,
and go to the device tab. The fields should be blank. Log out and
log back in, and the fields in the device tab should be populated,
because chrome restarted and thus fetched the properties again. With
this fix in place, the fields should be populated first time, every
time.

R=nkostylev@chromium.org,zel@chromium.org

Review URL: http://codereview.chromium.org/6903080

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@84926 4ff67af0-8c30-449e-8e8b-ad334ec8d88c