Fix hover state in wrench menu
Hiding the bookmark bar from the wrench menu would cause the wrench menu to get stuck with a hover state.

The problem was that when the super view is resized we don't get a mouse exited event in the button cell.

In this change I'm working around this by re-checking if the mouse is in the view once the wrench menu is closed. An alternative would be to use NSTrackingArea to get mouse entered/exited events. I think using NSTrackingArea would be a lot more work but if people would like me to use that I'll update this change to use that.

BUG=82456
TEST=Ran and verified that the hover state was no longer stuck.


Review URL: http://codereview.chromium.org/6969058

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@85427 4ff67af0-8c30-449e-8e8b-ad334ec8d88c