Refactor methods that enable and disable sync in the integration tests
The sync integration tests that completely disable sync fail on the
chromium buildbots, but are green on the trybots. Unfortunately, the
methods that disable / enable sync do not help in revealing why sync
failed.

This patch restructures those methods to return a boolean result that
can be verified by tests, and to do a better job of logging the state
of the syncer in case of a failure.

This is a first step in identifying the root cause of crbug.com/82489 on
the sync buildbots.

BUG=82489
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/7011050

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@85503 4ff67af0-8c30-449e-8e8b-ad334ec8d88c