Hide translate infobar on programmatic navigation.
In Google news, you can change the UI language through a popup menu [which does so by submitting a form programmatically].  The translation infobar wasn't being refreshed in this situation because the page navigation being performed is programmatic.

Infobar hiding on navigation, goes through TabContents::ExpireInfoBars() which has an early return if the navigation just performed wasn't initiated by the user.  After this, the code loops through the list of infobars and calls ShouldExpire() on each one to determine if it can be closed.

This CL removes the early return and moves the code to determine dismissal to the ShouldExpire() method of the infobar delegates, thus allowing the translate infobar to indicate that it can be dismissed on page navigation.

BUG=70261
TEST=See bug.

Review URL: http://codereview.chromium.org/6883299

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@85731 4ff67af0-8c30-449e-8e8b-ad334ec8d88c