Improve HQP Title Scoring
I was not satisfied with how difficult it was to get reasonable scores for quite long page titles even when multiple words contained in the title were typed so I added a limit on how much of the title (and URL) string was required in order to have a decent match.

BUG=None
TITLE=Ran all unit tests.
Review URL: http://codereview.chromium.org/7040009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@86192 4ff67af0-8c30-449e-8e8b-ad334ec8d88c