Take out pref watching from HostZoomMap, in preparation of removing Profile from it.
I've moved watching changes to the default zoom level to the Profile class.  The other value being watched was for the per-hostname level, which is only needed if it's being synced.  However this pref stopped being synced in r51849, so the removal is roughly a revert of 44056.
Review URL: http://codereview.chromium.org/7058003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@86303 4ff67af0-8c30-449e-8e8b-ad334ec8d88c