ntp4: Make theme attribution separate from chrome logo.
now the chrome logo should always show, but the attribution only shows if the theme has one. I also moved the attribution to the right side because it looks more balanced over there (this didn't break horribly for any of the handful of themes I tested on). Leaving it on the left side, right next to the chrome logo, looks cluttered.

BUG=none
TEST=manual

Review URL: http://codereview.chromium.org/6966022

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@86363 4ff67af0-8c30-449e-8e8b-ad334ec8d88c