Change the web store private install API to accept a localized extension name.
In our initial design of the beginInstallWithManifest function, we forgot that
some extensions have a name which needs token substitution for i18n/l10n. This
change renames beginInstallWithManifest to beginInstallWithManifest2, and
modifies the signature to take optional localizedName and locale parameters.

I also refactored how we compare the passed in manifest used in the
pre-download confirmation dialog - we now keep a copy of the source from the
.crx file to compare against, since i18n substututions can happen on the
manifest we use to construct the Extension object.

BUG=75821
TEST=Covered by browser tests for now; will require web store server-side
changes before it can be manually tested.

Review URL: http://codereview.chromium.org/6992047

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@86780 4ff67af0-8c30-449e-8e8b-ad334ec8d88c