CrOS - Send TPM PIN to flimflam for 802.1X Wi-Fi connections.
Some enterprise 802.1X Wi-Fi connection methods require user certificates, which are stored in the TPM.  The TPM PIN is "well known" but flimflam doesn't know it - so we forward it along.  We'll need to do something similar in the future for VPN connections that require user certificates.

BUG=chromium-os:15585
TEST=Verify that you can connect to an EAP-TLS network with a user certificate in the TPM.

Review URL: http://codereview.chromium.org/7062022

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@86854 4ff67af0-8c30-449e-8e8b-ad334ec8d88c